Implement partially redacted fields
The objective of this feature is to allow for providing a list of
fields which should be only partially censored. The primary case
for this is fields like ObjectId, which is often a URL to a file.
We may wish to hide the filename for certain security scenarios,
but still retain the ability to see what SharePoint site or
Document Library was the object of a given operation.