Implement self-contained MSDeploy target
End-Of-Day commit

Create a small webjob for processing the audit event queue,
attached to a Suave-based webapp.

Have Travis CI package the MSDeploy and push it to a new tag,
then create a release on GitHub and a publicly available file on
Azure Blob storage for reference in the azuredeploy.json ARM
template (which is wip.)

Also consolidate some dependencies, fix some compiler warnings,
binding redirects for Azure Web App env, etc etc.